Added transpiler test case and fixed bug

+++ python/common/python/math.java
@@ -14,9 +14,9 @@ public class math extends org.python.types.Module {
             "\n",
         args = {"x"}
     )
-    public static double ceil(org.python.Object x) {
+    public static int ceil(org.python.Object x) {
         double d =((org.python.types.Float) x.__float__()).value;
-        return Math.ceil(d);//Math.ceil(d);
+        return (int) Math.ceil(d);//Math.ceil(d);
         //throw new org.python.exceptions.NotImplementedError("Ceil not implemented yet.");
     }
